@@ -41,7 +41,7 @@ dependencies:
   string_2_icon: ^0.0.2
   flutter_rating_bar: ^4.0.1
   file_picker: ^8.0.3
-  get: ^4.6.1 
+  get: ^4.6.1
   video_player: ^2.2.11
   path_provider: ^2.0.2
   file: ^7.0.0
@@ -50,11 +50,12 @@ dependencies:
   flutter_custom_clippers: ^2.1.0
 
   #pusher_client: ^2.0.0
-  pusher_channels_flutter: ^1.0.0
-  intl: ^0.18.0 
-  web_socket_channel: ^2.1.0
+  pusher_channels_flutter: ^2.2.1
+  intl: ^0.19.0
+  web_socket_channel: ^3.0.0
 
-  
+
+
 dev_dependencies:
   flutter_test:
     sdk: flutter
@@ -64,7 +65,7 @@ dev_dependencies:
   # activated in the `analysis_options.yaml` file located at the root of your
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
-  flutter_lints: ^3.0.0
+  flutter_lints: ^4.0.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
@@ -91,6 +92,7 @@ flutter:
      - assets/getjob.png
      - assets/getservice.png
      - assets/getcourse.png
+     - assets/book.webp
 
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware
